Update ObbExpansionsManager.java
Hi Sergey.  I have modified this code to resolve the failure to invoke the OnObbStateChanged callback.
The problem was that the StorageManager only holds weak references to the listeners.  When the  program block that created the listener finished the reference was being freed and subsequently nulled in the StorageManager, all before the mount completed.  

By making the listeners into instance variables the reference is maintained and the problem is avoided.

I believe that with this change the MountChecker is no longer required, but have left it in place with a substantially extended initial call.  See what you think.  

I found the ObbExpansionsManager to be an incredibly useful piece of code which gave me an excellent introduction to mountable obb files and how to manage them;  thank you so much for sharing this.  

Regards, Paul.